docs: Improves documentation on the refactored code
