+++ src/on-chain-data-client/airlines/airline.js
@@ -64,6 +64,13 @@ class OnChainAirline extends OnChainRecord implements AirlineInterface {
     return this.indexContract.methods.deleteAirline(this.address);
   }
 
+  /**
+   * Generates transaction data and metadata for creating new airline contract on-chain.
+   * Transaction is not signed nor sent here.
+   *
+   * @param {TransactionOptionsInterface} options object, only `from` property is currently used, all others are ignored in this implementation
+   * @return {Promise<PreparedTransactionMetadataInterface>} Transaction data and metadata, including the freshly created airline instance.
+   */
   async createOnChainData (transactionOptions: TransactionOptionsInterface): Promise<PreparedTransactionMetadataInterface> {
     const result = await this._createOnChainData(transactionOptions);
     result.airline = result.record;
@@ -71,6 +78,19 @@ class OnChainAirline extends OnChainRecord implements AirlineInterface {
     return result;
   }
 
+  /**
+   * This is potentially devastating, so it's better to name
+   * this operation explicitly instead of hiding it under updateOnChainData.
+   *
+   * Generates transaction data and metadata required for a airline ownership
+   * transfer.
+   *
+   * @param {string} Address of a new manager
+   * @param {TransactionOptionsInterface} options object, only `from` property is currently used, all others are ignored in this implementation
+   * @throws {SmartContractInstantiationError} When the underlying contract is not yet deployed.
+   * @return {Promise<PreparedTransactionMetadataInterface>} Transaction data and metadata, including the freshly created airline instance.
+   *
+   */
   async transferOnChainOwnership (newManager: string, transactionOptions: TransactionOptionsInterface): Promise<PreparedTransactionMetadataInterface> {
     const result = await this._transferOnChainOwnership(newManager, transactionOptions);
     result.airline = result.record;
@@ -78,6 +98,13 @@ class OnChainAirline extends OnChainRecord implements AirlineInterface {
     return result;
   }
 
+  /**
+   * Generates transaction data and metadata required for destroying the airline object on network.
+   *
+   * @param {TransactionOptionsInterface} options object, only `from` property is currently used, all others are ignored in this implementation
+   * @throws {SmartContractInstantiationError} When the underlying contract is not yet deployed.
+   * @return {Promise<PreparedTransactionMetadataInterface>} Transaction data and metadata, including the freshly created airline instance.
+   */
   async removeOnChainData (transactionOptions: TransactionOptionsInterface): Promise<PreparedTransactionMetadataInterface> {
     const result = await this._removeOnChainData(transactionOptions);
     result.airline = result.record;
@@ -85,6 +112,15 @@ class OnChainAirline extends OnChainRecord implements AirlineInterface {
     return result;
   }
 
+  /**
+   * Generates transaction data and metadata required for all airline-related data modification
+   * by calling `updateRemoteData` on a `RemotelyBackedDataset`.
+   *
+   * @param {TransactionOptionsInterface} options object that is passed to all remote data setters
+   * @throws {SmartContractInstantiationError} When the underlying contract is not yet deployed.
+   * @throws {SmartContractInstantiationError} When dataUri is empty.
+   * @return {Promise<Array<PreparedTransactionMetadataInterface>>} List of transaction metadata
+   */
   async updateOnChainData (transactionOptions: TransactionOptionsInterface): Promise<Array<BasePreparedTransactionMetadataInterface>> {
     const results = (await this._updateOnChainData(transactionOptions))
       .map((result) => {
